v1.9 Basic functionality is working. Final touches need to be done.

+++ includes/src/internal/competition.rounds.php
@@ -14,7 +14,8 @@ class round
     var $id;
     var $db;
 
-    public function __construct($id = null){
+    public function __construct($id = null)
+    {
         $this->id = $id;
         //echo __DIR__;
 
@@ -28,29 +29,64 @@ class round
                 'username' => \svn\settings::dbUsername,
                 'password' => \svn\settings::dbPassword)
         );
+
     }
 
     /**
      * @param $competitionId
-     * @param null $roundId
-     *
-     * Query the existing round + providing details
+     * @param null $round
      * @return array
+     * @internal param null $roundId Query the existing round + providing details*
+     * Query the existing round + providing details
      */
-    public function getRounds($competitionId){
-        $data = $this->db->select('svn_rounds', '*', array('comp_id' => $competitionId));
+    public function getRounds($competitionId, $round = false)
+    {
+        if ($competitionId && !$round)
+            $data = $this->db->select('svn_rounds', '*', array('comp_id' => $competitionId));
+        else
+            $data = $this->db->select('svn_rounds', '*', array('id' => $round));
+        foreach ($data as $key => $row) {
+            $data[$key]["matrix_score"] = unserialize($data[$key]["matrix_score"]);
+            $data[$key]["matrix_games"] = unserialize($data[$key]["matrix_games"]);
+            $data[$key]["matrix_win"] = unserialize($data[$key]["matrix_win"]);
+            $data[$key]["matrix_draw"] = unserialize($data[$key]["matrix_draw"]);
+            $data[$key]["matrix_loss"] = unserialize($data[$key]["matrix_loss"]);
+        }
         return $data;
     }
 
     /**
      * @param $competitionId
-     * @param $roundData Array(compId, round, date)
+     * @param $round
+     * @return bool
+     *
+     * Returns the id of a round based upon competitionId and round number
+     */
+    public function getRoundId($competitionId, $round)
+    {
+        if ($round && $competitionId) {
+            $data = $this->db->select("svn_rounds", "id", array("AND" => array("comp_id" => $competitionId, "round" => $round)));
+            $this->id = $data[0];
+            return $this->id;
+        }
+        return false;
+    }
+
+    /**
+     * @param $competitionId
+     * @param $round Array(compId, round, date)
      *
      * Create a new round
      */
-    public function createRound($roundData){
-        $data = $this->db->insert('svn_rounds', $roundData);
-        return $data;
+    public function createRound($round)
+    {
+        //Get highest round
+        $maxRound = $this->db->max('svn_rounds', 'round', array('comp_id' => $round->comp_id));
+        $round->round = $maxRound + 1;
+        //Create the round
+        $data = $this->db->insert('svn_rounds', array("comp_id" => $round->comp_id, "round" => $round->round, "date" => $round->date));
+        $data = $this->getRounds(null, $data);
+        return $data[0];
     }
 
     /**
@@ -59,8 +95,9 @@ class round
      *
      * Update existing round
      */
-    public function updateRound($roundData){
-        if($this->id) {
+    public function updateRound($roundData)
+    {
+        if ($this->id) {
             $data = $this->db->update('svn_rounds', $roundData, array("id" => $this->id));
             return $data;
         }
@@ -69,15 +106,32 @@ class round
     /**
      * @return bool|int
      */
-    public function deleteRound(){
-        if($this->id) {
+    public function deleteRound()
+    {
+        if ($this->id) {
             // Delete games
+            $data = $this->db->delete('svn_partijen', array("round_id" => $this->id));
+            if ($data !== false) {
 
-            // Delete standing
+                // Delete standing
 
-            // Delete round
-            $data = $this->db->delete('svn_rounds', array("id" => $this->id));
-            return $data;
+                // Delete round
+                $data = $this->db->delete('svn_rounds', array("id" => $this->id));
+                return $data;
+            }
         }
+    }
+
+    /**
+     * @param $competitionId
+     * @return bool|int|string
+     *
+     * Gets the round nr of the previous round
+     */
+    public function getPreviousRound($competitionId, $round)
+    {
+        if ($round > 1)
+            return $this->db->max('svn_rounds', 'round', array("AND" => array("comp_id" => $competitionId, "round[<]" => $round)));
+        return 0;
     }
 }
